+++ ecommerce-react/src/components/Footer/Footer.css
@@ -11,7 +11,7 @@
       padding-top: 20px;
   }
 
-  .icon-btn {
+  .icon-btn-style {
       color: white;
   }
