+++ ecommerce-react/src/components/Item/Item.css
@@ -12,5 +12,5 @@
 }
 
 .cardItem{
-    width: 270px;
+    width: 85%;
 }
