+++ ecommerce-react/src/pages/Home.js
@@ -1,14 +1,17 @@
 import React from 'react'
 import './HomePage.css'
+import ItemListContainer from '../components/ItemListContainer/ItemListContainer';
 
-function HomePage(){
+function HomePage({products}){
   return(
-      <div className='back-home'>
-          <div className='div-home'>
-              <p>¡Bienvenid@!</p>
-              <p>Accedé a las mejores promociones y precios.</p>
-          </div>
-      </div>
+      <>
+        <div className='back-home'>
+            <div className='div-home'>
+                <p>¡Bienvenid@!</p>
+            </div>
+        </div>
+        <ItemListContainer title={'This is a test'} allProducts={products}/>
+      </>
   );
 }
