+++ ecommerce-react/src/pages/HomePage.css
@@ -22,7 +22,7 @@
 }
 
 .div-home p{
-    font-size: 100px;
+    font-size: 50px;
     text-align: center;
     color: white;
 }
